Add in edits from Ben Craig, and more.
Ben provided some really important edits to the motivation section:
clear indication of interoperation as the main driver. With a tweak to
the first paragraph and the addition of an interop paragraph.

Avoided the use of "codify" in the scope. And added the points
from Ben on what is out of scope. With a couple of explanations to
go with them.